Move delete item to another file
