+++ src/index.js
@@ -4,12 +4,11 @@ import {
   newTaskForm,
   addTask,
   clearCompletedTasksButton,
+  deleteItem,
 } from './addRemove.js';
 import { checkCompleted } from './interactive.js';
 import removeAllCheckedRender from './removeAll.js';
-import { tasks, save } from './localStorage.js';
-
-window.addEventListener('DOMContentLoaded', render(tasks));
+import { tasks } from './localStorage.js';
 
 newTaskForm.addEventListener('submit', addTask);
 
@@ -21,16 +20,9 @@ tasksContainer.addEventListener('click', (e) => {
 
 clearCompletedTasksButton.addEventListener('click', removeAllCheckedRender);
 
-const listItems = document.querySelectorAll('li');
-listItems.forEach((item, index) => {
-  item.addEventListener('click', (e) => {
-    const deleteBtn = e.target;
-    if (deleteBtn.classList[0] === 'far') {
-      let tasks = JSON.parse(localStorage.getItem('tasks')) || [];
-      tasks = tasks.filter((task) => task !== tasks[index]);
-      item.remove();
-      save(tasks);
-      render(tasks);
-    }
-  });
+window.addEventListener('DOMContentLoaded', () => {
+  render(tasks);
+  deleteItem();
+  addTask();
+  removeAllCheckedRender();
 });
